- item endpoints now take String username (instead of int userId)

- import lines cleaned in all item java files

- header comment added all item java files

+++ src/main/java/com/wishlist/cst438project2/controller/AdminController.java
@@ -1,19 +1,21 @@
 package com.wishlist.cst438project2.controller;
 
 import com.wishlist.cst438project2.common.Constants;
-import com.wishlist.cst438project2.dto.ItemDTO;
-import com.wishlist.cst438project2.dto.SignInDTO;
-import com.wishlist.cst438project2.dto.SignUpDTO;
-import com.wishlist.cst438project2.dto.UserDTO;
+import com.wishlist.cst438project2.common.TokenManager;
+import com.wishlist.cst438project2.dto.*;
+import com.wishlist.cst438project2.enums.RoleType;
 import com.wishlist.cst438project2.exception.BadRequestException;
+import com.wishlist.cst438project2.exception.UnauthorizedException;
 import com.wishlist.cst438project2.service.AdminService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 import java.util.Objects;
 
+@CrossOrigin
 @RestController
 @RequestMapping("/v1/admin")
 @Slf4j
@@ -22,12 +24,20 @@ public class AdminController {
     @Autowired
     private AdminService adminService;
 
+    @Autowired
+    private TokenManager tokenManager;
+
     @GetMapping("/users")
-    public List<UserDTO> getAllUsers() {
+    public List<UserDTO> getAllUsers(@RequestHeader String accessToken) {
 
         log.info("AdminController: Starting getAllUsers");
 
         try {
+
+            UserTokenDTO userTokenDTO = tokenManager.getUser(accessToken);
+
+            if(Objects.isNull(userTokenDTO) || !userTokenDTO.getRole().equals(RoleType.ADMIN.getValue()))
+                throw new UnauthorizedException(Constants.ERROR_INVALID_TOKEN);
 
             List<UserDTO> userDTOList = adminService.getAllUsers();
 
@@ -46,11 +56,15 @@ public class AdminController {
      * @param username, of the user whose account is to be deleted.
      */
     @DeleteMapping("/deleteUser")
-    public String deleteUser(@RequestParam String username) {
+    public String deleteUser(@RequestHeader String accessToken, @RequestParam String username) {
 
         log.info("AdminController: Starting deleteUser");
 
         try{
+            UserTokenDTO userTokenDTO = tokenManager.getUser(accessToken);
+
+            if(Objects.isNull(userTokenDTO) || !userTokenDTO.getRole().equals(RoleType.ADMIN.getValue()))
+                throw new UnauthorizedException(Constants.ERROR_INVALID_TOKEN);
 
             if(Objects.isNull(username) || username.isEmpty())
                 throw new BadRequestException();
@@ -73,11 +87,15 @@ public class AdminController {
      * @return user creation timestamp
      */
     @PostMapping("/createUser")
-    public String createUser(@RequestBody SignUpDTO signUpDTO) {
+    public String createUser(@RequestHeader String accessToken, @RequestBody SignUpDTO signUpDTO) {
 
         log.info("AdminController: Starting createUser");
 
         try {
+            UserTokenDTO userTokenDTO = tokenManager.getUser(accessToken);
+
+            if(Objects.isNull(userTokenDTO) || !userTokenDTO.getRole().equals(RoleType.ADMIN.getValue()))
+                throw new UnauthorizedException(Constants.ERROR_INVALID_TOKEN);
 
             if (Objects.isNull(signUpDTO))
                 throw new BadRequestException();
@@ -95,25 +113,36 @@ public class AdminController {
     }
 
     @PostMapping("/login")
-    public String login(@RequestBody SignInDTO signInDTO) {
+    public ResponseDTO<UserLoginDTO> login(@RequestBody SignInDTO signInDTO) {
 
         log.info("AdminController: Starting login");
 
+        ResponseDTO<UserLoginDTO> responseDTO = new ResponseDTO<>();
+        UserLoginDTO userLoginDTO = null;
+        String message = HttpStatus.UNAUTHORIZED.toString();
+        int httpStatusCode;
         try {
 
             if (Objects.isNull(signInDTO))
                 throw new BadRequestException();
 
-            String responseTimestamp = adminService.login(signInDTO);
+            userLoginDTO = adminService.login(signInDTO);
+            if(Objects.isNull(userLoginDTO))
+                throw new UnauthorizedException();
 
+            httpStatusCode = HttpStatus.OK.value();
+            message = Constants.USER_LOGIN_SUCCESSFUL;
             log.info("AdminController: Exiting login");
-
-            return responseTimestamp;
 
         } catch (Exception ex) {
             log.error(ex.getMessage(), ex);
-            throw ex;
+            httpStatusCode = HttpStatus.UNAUTHORIZED.value();
         }
+
+        responseDTO.setStatus(httpStatusCode);
+        responseDTO.setData(userLoginDTO);
+        responseDTO.setMessage(message);
+        return responseDTO;
     }
 
     /**
@@ -121,9 +150,14 @@ public class AdminController {
      * returns item creation timestamp
      */
     @PostMapping("/createItem")
-    public String createItem(@RequestBody ItemDTO itemDTO) {
+    public String createItem(@RequestHeader String accessToken, @RequestBody ItemDTO itemDTO) {
         log.info("ItemController: Starting createItem");
         try {
+            UserTokenDTO userTokenDTO = tokenManager.getUser(accessToken);
+
+            if(Objects.isNull(userTokenDTO) || !userTokenDTO.getRole().equals(RoleType.ADMIN.getValue()))
+                throw new UnauthorizedException(Constants.ERROR_INVALID_TOKEN);
+
             if (Objects.isNull(itemDTO) || (itemDTO.getName().isBlank() || Objects.isNull(itemDTO.getUserId()))) {
                 throw new BadRequestException();
             } else {
@@ -142,9 +176,47 @@ public class AdminController {
      * returns timestamp of successful deletion
      */
     @DeleteMapping("/removeItems")
-    public String removeItem(@RequestParam String itemName, @RequestParam int userId) {
+    public String removeItem(@RequestHeader String accessToken, @RequestParam String itemName, @RequestParam int userId) {
+
+        UserTokenDTO userTokenDTO = tokenManager.getUser(accessToken);
+
+        if(Objects.isNull(userTokenDTO) || !userTokenDTO.getRole().equals(RoleType.ADMIN.getValue()))
+            throw new UnauthorizedException(Constants.ERROR_INVALID_TOKEN);
+
         log.info("ItemController: Starting removeItem");
         log.info("ItemController: removeItem:\n    name: {}\n    userId: {}", itemName, userId);
         return adminService.removeItem(itemName, userId);
+    }
+
+    /**
+     * This API is used for updating user information in database
+     * @param userDTO, whose details is to be updated
+     * @return user creation timestamp
+     */
+    @PutMapping("/updateUser")
+    public UserDTO updateUser(@RequestHeader String accessToken, @RequestBody UserDTO userDTO) {
+
+        log.info("UserController: Starting updateUser");
+
+        try {
+
+            UserTokenDTO userTokenDTO = tokenManager.getUser(accessToken);
+
+            if(Objects.isNull(userTokenDTO) || !userTokenDTO.getRole().equals(RoleType.ADMIN.getValue()))
+                throw new UnauthorizedException(Constants.ERROR_INVALID_TOKEN);
+
+            if(Objects.isNull(userDTO))
+                throw new BadRequestException();
+
+            userDTO = adminService.updateUser(userDTO);
+
+            log.info("UserController: Exiting updateUser");
+
+            return userDTO;
+
+        } catch (Exception ex) {
+            log.error(ex.getMessage(), ex);
+            throw ex;
+        }
     }
 }
